api/requirements.txt
flask
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
gunicorn
Flask-Migrate
psycopg2
pytest
pytest-flask
pytest-mock
pytest-cov
black
